Reorganizes code
Basic code cleanup, tidying the directories for less confusion
